% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eoPredData-package.R
\docType{package}
\name{eoPredData-package}
\alias{eoPredData}
\alias{eoPredData-package}
\title{eoPredData: ExperimentHub package for eoPred}
\description{
Provides access to eoPred pretrained model hosted on ExperimentHub. Model was trained on placental DNA methylation preeclampsia samples using mixOmics splsda. There are two resources: 1. the model object, and 2. a testing data set used to demonstrate the function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wvictor14/eoPredData}
  \item Report bugs at \url{https://github.com/wvictor14/eoPredData/issues}
}

}
\author{
\strong{Maintainer}: Victor Yuan \email{victor.2wy@gmail.com}

Authors:
\itemize{
  \item Iciar Boyano \email{Iciar.Fernandez@bcchr.ca}
}

}
\keyword{internal}
